Move spellsTab logic from MainBattleController to separate SpellsTab class
